+++ lib/app/router/app_route.dart
@@ -1,7 +1,8 @@
 enum AppRoute {
   splash(path: '/', name: 'splash'),
   auth(path: '/auth', name: 'auth'),
-  hotDesk(path: '/desk', name: 'hotDesk');
+  hotDesk(path: '/desk', name: 'hotDesk'),
+  admin(path: '/admin', name: 'admin');
 
   const AppRoute({
     required this.path,